Refactor components into modular files
