+++ public/js/components/index.js
@@ -1,0 +1,5 @@
+export * from './elements.js';
+export * from './text.js';
+export * from './inputs.js';
+export * from './media.js';
+export * from './controls.js';
